[benc/numeric-input-refactor] Move functions to utils.ts
