+++ packages/perseus/src/widgets/numeric-input/numeric-input.tsx
@@ -1,4 +1,3 @@
-import {KhanMath} from "@khanacademy/kmath";
 import {useOnMountEffect} from "@khanacademy/wonder-blocks-core";
 import * as React from "react";
 import {
@@ -19,6 +18,10 @@ import styles from "./numeric-input.module.css";
 import stylesLegacy from "./numeric-input_legacy-styles";
 import {
     generateExamples,
+    getCorrectUserInput,
+    getOneCorrectAnswerFromRubric,
+    getStartUserInput,
+    getUserInputFromSerializedState,
     normalizeCorrectAnswerForms,
     shouldShowExamples,
 } from "./utils";
@@ -26,8 +29,6 @@ import {
 import type {Focusable, Widget, WidgetExports, WidgetProps} from "../../types";
 import type {NumericInputPromptJSON} from "../../widget-ai-utils/numeric-input/prompt-utils";
 import type {
-    MathFormat,
-    PerseusNumericInputRubric,
     PerseusNumericInputUserInput,
     PerseusNumericInputWidgetOptions,
 } from "@khanacademy/perseus-core";
@@ -202,111 +203,6 @@ export const NumericInput = forwardRef<Widget, Props>(
         );
     },
 );
-
-function getStartUserInput(): PerseusNumericInputUserInput {
-    return {currentValue: ""};
-}
-
-/**
- * @deprecated and likely a very broken API
- * [LEMS-3185] do not trust serializedState
- */
-function getUserInputFromSerializedState(
-    serializedState: any,
-): PerseusNumericInputUserInput {
-    return {
-        currentValue: serializedState.currentValue,
-    };
-}
-
-export function findPrecision(value: number) {
-    for (let i = 0; i < 10; i++) {
-        // `toFixed` handily rounds a number to a given precision...
-        // ...but also turns it into a string. so `+` turns it back
-        // into a number.
-        if (value === +value.toFixed(i)) {
-            return i;
-        }
-    }
-    return 10; // don't assume there's more precision than that
-}
-
-export function findCommonFractions(value: number) {
-    const whole = Math.floor(value);
-    if (value === whole) {
-        return;
-    }
-    const decimal = value - whole;
-    const precision = findPrecision(decimal);
-    // it's brute force, but it's honest work
-    for (let num = 1; num < 100; num++) {
-        for (let denom = 2; denom < 100; denom++) {
-            if (+(num / denom).toFixed(precision) === decimal) {
-                return {num: num + whole * denom, denom};
-            }
-        }
-    }
-}
-
-function getCorrectUserInput(
-    options: PerseusNumericInputWidgetOptions,
-): PerseusNumericInputUserInput {
-    for (const answer of options.answers) {
-        if (answer.status === "correct" && answer.value != null) {
-            if (answer.answerForms?.includes("decimal")) {
-                return {currentValue: answer.value.toString()};
-            }
-            if (answer.answerForms?.includes("improper")) {
-                const frac = findCommonFractions(answer.value);
-                if (frac) {
-                    return {currentValue: `${frac.num}/${frac.denom}`};
-                }
-            }
-            if (answer.answerForms?.includes("proper")) {
-                const frac = findCommonFractions(answer.value);
-                if (frac) {
-                    const {num, denom} = frac;
-                    if (num > denom) {
-                        const whole = Math.floor(num / denom);
-                        const remainder = num - whole * denom;
-                        return {currentValue: `${whole} ${remainder}/${denom}`};
-                    } else {
-                        return {currentValue: `${num}/${denom}`};
-                    }
-                }
-            }
-            // 🤷
-            return {currentValue: answer.value.toString()};
-        }
-    }
-    return {currentValue: ""};
-}
-
-function getOneCorrectAnswerFromRubric(
-    rubric: PerseusNumericInputRubric,
-): string | null | undefined {
-    const correctAnswers = rubric.answers.filter(
-        (answer) => answer.status === "correct",
-    );
-    const answerStrings = correctAnswers.map((answer) => {
-        // Either get the first answer form or default to decimal
-        const format: MathFormat =
-            answer.answerForms && answer.answerForms[0]
-                ? answer.answerForms[0]
-                : "decimal";
-
-        let answerString = KhanMath.toNumericString(answer.value!, format);
-        if (answer.maxError) {
-            answerString +=
-                " ± " + KhanMath.toNumericString(answer.maxError, format);
-        }
-        return answerString;
-    });
-    if (answerStrings.length === 0) {
-        return;
-    }
-    return answerStrings[0];
-}
 
 export default {
     name: "numeric-input",
